% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.requested.check.R
\name{analyses.check}
\alias{analyses.check}
\title{Ensure Analyses Request Vector is Well-formed}
\usage{
analyses.check(analyses.requested)
}
\arguments{
\item{analyses.requested}{A named logical vector with elements named Global, Locus, and Pairwise}
}
\description{
take a named vector and make sure that it contains the names Global, Locus, and Pairwise
}

