% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funcs.R
\name{analysis_func}
\alias{analysis_func}
\title{Analysis functions}
\usage{
analysis_func(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\description{
Take results of a simulation from params@rep.sample and run
  Global, Locus, and Pairwise analyses
}

