% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funcs.R
\name{analysis_funcs}
\alias{analysis_funcs}
\alias{calcChordDist}
\alias{globalAnalysis}
\alias{hapSmryFunc}
\alias{loadResultsMatrix}
\alias{locusAnalysisGenotypes}
\alias{locusAnalysisHaplotypes}
\alias{overall_stats}
\alias{pairwiseAnalysis}
\title{Analysis functions}
\usage{
analysis_funcs(params)

loadResultsMatrix(params, mat, label)

overall_stats(g, num.perm.reps, num.cores)

globalAnalysis(g, num.perm.reps, num.cores)

locusAnalysisGenotypes(g)

hapSmryFunc(g)

locusAnalysisHaplotypes(g)

calcChordDist(dat, is.diploid)

pairwiseAnalysis(g, num.perm.reps, num.cores)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}

\item{mat}{results matrix to be loaded into params object.}

\item{label}{analysis type label ("Global", "Locus", or "Pairwise").}

\item{g}{a gtypes object.}

\item{num.perm.reps}{number of permutation replicates.}

\item{num.cores}{number of CPU cores to use.}

\item{dat}{data.frame in hierfstat format (see \code{\link[hierfstat]{genet.dist}}).}

\item{is.diploid}{logical - is this a diploid object?}
}
\description{
Run Global, Locus, and Pairwise analyses on results from
  a single simulation replicate stored in params@rep.sample#'
}

