% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMratio.R
\name{calc.mratio}
\alias{calc.mratio}
\title{Garza Williamson M Ratio (bottleneck) Statistic, Per Pop, Per Loc}
\usage{
calc.mratio(gen.data.gtype)
}
\arguments{
\item{gen.data.gtype}{a gtype object}
}
\description{
calculates the mratio from a gtypes object
}

