% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.classes.R
\docType{class}
\name{fastsimcoal.params-class}
\alias{fastsimcoal.params-class}
\title{fastsimcoal Parameters Class}
\description{
An S4 class storing parameters specific to fastsimcoal
}
\section{Slots}{

\describe{
\item{\code{fastsimcoal.exec}}{character string for the fastsimcoal command line
executable.}

\item{\code{sample.times}}{a vector giving the number of generations in the past
at which samples are taken.}

\item{\code{growth.rate}}{a vector giving the growth rate of each population.}

\item{\code{hist.ev}}{a matrix describing historical events.}

\item{\code{num.chrom}}{a value giving the number of chromosomes that the
\code{locus.params} marker specifications should be copied for. If
\code{NULL}, then chromosome assignment is taken from the
\code{chromosome} column in \code{locus.params}. Any non-\code{NULL}
integer will cause the \code{chromosome} column to be ignored.}

\item{\code{locus.params}}{a data.frame giving the parameters for each locus.}
}}

