% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.histEvMat.R
\name{fsc.histEvMat}
\alias{fsc.histEvCheck}
\alias{fsc.histEvConverges}
\alias{fsc.histEvMat}
\title{Create fastsimcoal historical event matrices}
\usage{
fsc.histEvMat(num.events = 0)

fsc.histEvConverges(hist.ev, pop.size, growth.rate, num.mig.mats = NULL)

fsc.histEvCheck(hist.ev, pop.size, growth.rate, num.mig.mats = NULL)
}
\arguments{
\item{num.events}{number of historical events.}

\item{hist.ev}{a matrix describing historical events, with one row per event.}

\item{pop.size}{numerical vector giving size of each population.}

\item{growth.rate}{numerical vector giving growth rate of each population.}

\item{num.mig.mats}{number of migration matrices.}
}
\value{
a blank fastsimcoal historical event matrices that can be
  filled in later
}
\description{
Create fastsimcoal historical event matrices
}

