% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.loadScenario.R
\name{fsc.loadScenario}
\alias{fsc.loadScenario}
\title{Load fastsimcoal locus parameters}
\usage{
fsc.loadScenario(num.pops, pop.size, sample.size, migration, mut.rate,
  sample.times = NULL, growth.rate = NULL, hist.ev = NULL,
  locus.type = c("dna", "msat", "snp"), sequence.length = NULL,
  num.loci = NULL, transition.rate = NULL, gsm.param = NULL,
  range.constraint = NULL, min.freq = NULL, recomb.rate = NULL,
  chromosome = NULL, num.chrom = NULL)
}
\arguments{
\item{num.pops}{number of populations.}

\item{pop.size}{a vector \code{num.pop} long giving size of each populaiton.}

\item{sample.size}{a vector \code{num.pop} long giving the number of
samples to take from each population.}

\item{migration}{a \code{num.pop} x \code{num.pop} matrix giving the
migration rates between each population.}

\item{mut.rate}{mutation rate for DNA or msat.}

\item{sample.times}{a vector giving the number of generations in the past
at which samples are taken.}

\item{growth.rate}{a vector giving the growth rate of each population.}

\item{hist.ev}{a matrix describing historical events.}

\item{locus.type}{a character representation of what type of marker to simulate.
Can be "dna", "msat", or "snp".}

\item{sequence.length}{number of DNA base pairs to use.}

\item{num.loci}{number of msat or snp loci to simulate.}

\item{transition.rate}{DNA: fraction of substitutions that are transitions.}

\item{gsm.param}{Value of the geometric parameter for a Generalized Stepwise
Mutation (GSM) model. This value represents the proportion of mutations
that will change the allele size by more than one step. Values between
0 and 1 are required. A value of 0 is for a strict
Stepwise Mutation Model (SMM).}

\item{range.constraint}{Range constraint (number of different alleles
allowed). A value of 0 means no range constraint}

\item{min.freq}{SNP: minimum frequency for the derived allele.}

\item{recomb.rate}{recombination rate between adjacent markers.}

\item{chromosome}{number or character identifying which chromosome the marker
is on.}

\item{num.chrom}{a value giving the number of chromosomes that the
\code{locus.params} marker specifications should be copied for. If
\code{NULL}, then chromosome assignment is taken from the
\code{chromosome} column in \code{locus.params}. Any non-\code{NULL}
integer will cause the \code{chromosome} column to be ignored.}
}
\description{
Load fastsimcoal locus parameters
}

