% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.accessors.R, R/fsc.locusParams.R
\docType{methods}
\name{loadLocusParams<-}
\alias{fsc.locus.dna}
\alias{fsc.locus.msat}
\alias{fsc.locus.snp}
\alias{fsc.locusParams}
\alias{loadLocusParams<-}
\alias{loadLocusParams<-,fastsimcoal.params-method}
\title{Create fastsimcoal locus parameter data.frames}
\usage{
loadLocusParams(x) <- value

\S4method{loadLocusParams}{fastsimcoal.params}(x) <- value

fsc.locus.dna(sequence.length, mut.rate, transition.rate = NULL,
  recomb.rate = NULL, chromosome = NULL)

fsc.locus.msat(num.loci, mut.rate, gsm.param = NULL,
  range.constraint = NULL, recomb.rate = NULL, chromosome = NULL)

fsc.locus.snp(num.loci, min.freq, recomb.rate = NULL, chromosome = NULL)
}
\arguments{
\item{x}{a \linkS4class{fastsimcoal.params} object.}

\item{value}{locus parameter data.frame to load.}

\item{sequence.length}{DNA: number of base pairs for sequence.}

\item{mut.rate}{DNA: mutation rate per bp, MICROSAT: mutation rate per locus.}

\item{transition.rate}{DNA: fraction of substitutions that are transitions.}

\item{recomb.rate}{recombination rate between adjacent markers.}

\item{chromosome}{number or character identifying which chromosome the marker
is on.}

\item{num.loci}{MICROSAT, SNP: number of loci to simulate.}

\item{gsm.param}{Value of the geometric parameter for a Generalized Stepwise
Mutation (GSM) model. This value represents the proportion of mutations
that will change the allele size by more than one step. Values between
0 and 1 are required. A value of 0 is for a strict
Stepwise Mutation Model (SMM).}

\item{range.constraint}{Range constraint (number of different alleles
allowed). A value of 0 means no range constraint}

\item{min.freq}{SNP: minimum frequency for the derived allele.}
}
\description{
Create fastsimcoal locus parameter data.frames
}

