% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.run.R
\name{fsc.run}
\alias{fsc.read}
\alias{fsc.run}
\alias{fsc.write}
\title{Run fastsimcoal}
\usage{
fsc.run(params)

fsc.write(num.pops, Ne, sample.size = NULL, sample.times = NULL,
  growth.rate = NULL, mig.rates = NULL, hist.ev = NULL,
  num.chrom = NULL, locus.params = NULL, ploidy = NULL, label = NULL)

fsc.read(file, chrom.pos, ploidy)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}

\item{num.pops}{number of populations.}

\item{Ne}{numeric vector: size of each population.}

\item{sample.size}{numeric vector: number of samples to draw from each population.}

\item{sample.times}{numeric vector: time step samples are drawn from for each population.}

\item{growth.rate}{numeric vector: growth rate of each population.}

\item{mig.rates}{list of matrices: migration rates between populations.}

\item{hist.ev}{matrix: one row per historical event.}

\item{num.chrom}{number of unique chromosomes.}

\item{locus.params}{data.frame: one row per locus parameter.}

\item{ploidy}{a numeric giving the ploidy of the loci (1 = haploid,
2 = diploid).}

\item{label}{character: label for file naming.}

\item{file}{character filename}

\item{chrom.pos}{a matrix giving the start and end positions of each chromosome.}
}
\value{
a modified \linkS4class{skeleSim.params} object with the results of
  a fastsimcoal run.
}
\description{
Run fastsimcoal
}

