% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.scenarioCheck.R
\name{fsc.scenarioCheck}
\alias{fsc.scenarioCheck}
\alias{fsc.histEvConverges}
\alias{fsc.histEvCheck}
\title{Check parameters for fastsimcoal}
\usage{
fsc.scenarioCheck(params)

fsc.histEvConverges(hist.ev, pop.size, growth.rate, num.mig.mats = NULL)

fsc.histEvCheck(hist.ev, pop.size, growth.rate, num.mig.mats = NULL)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}

\item{hist.ev}{a matrix describing historical events, with one row per event.}

\item{pop.size}{numerical vector giving size of each population.}

\item{growth.rate}{numerical vector giving growth rate of each population.}

\item{num.mig.mats}{number of migration matrices.}
}
\description{
Check parameters for fastsimcoal
}
