% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.run.R
\name{fsc.write}
\alias{fsc.write}
\title{Write fastsimcoal files}
\usage{
fsc.write(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\value{
a modified \linkS4class{skeleSim.params} object with the results of
  a fastsimcoal run.
}
\description{
Run fastsimcoal
}
