% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.make.genpop.R
\name{landscape.make.genpop}
\alias{landscape.make.genpop}
\title{Create genpop object from landscape}
\usage{
landscape.make.genpop(l)
}
\arguments{
\item{l}{a \code{rmetasim} landscape object.}
}
\description{
Create genpop object from landscape
}

