% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.all.stats.R
\name{plot.all.stats}
\alias{plot.all.stats}
\title{Plot all results}
\usage{
plot.all.stats(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\description{
Plot all results
}

