% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.convert.R
\name{rms.convert}
\alias{rms.convert}
\title{Extract genetic data from landscape object}
\usage{
rms.convert(Rland, locus.type)
}
\arguments{
\item{Rland}{a \code{rmetasim} landscape object.}

\item{locus.type}{character giving type of locus (\code{microsat} or
\code{sequence})}
}
\value{
A gtypes object containing genotypes (or sequences)
}
\description{
Extract genetic data from landscape object
}
