% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.init.land.function.R
\name{rms.init.landscape.func}
\alias{rms.init.landscape.func}
\title{Initialize a landscape object}
\usage{
rms.init.landscape.func(num.pops = NULL, carrying = NULL,
  sample.size = NULL, mig.rates = NULL, num.loc = NULL, loc.type = NULL,
  mut.rate = NULL, seq.length = NULL, num.stgs = NULL, selfing = NULL,
  surv.matr = NULL, repr.matr = NULL, male.matr = NULL,
  init.pop.sizes = NULL, num.gen = NULL, num.alleles = NULL,
  allele.freqs = NULL)
}
\arguments{
\item{num.pops}{?}

\item{carrying}{?}

\item{sample.size}{?}

\item{mig.rates}{?}

\item{num.loc}{?}

\item{loc.type}{?}

\item{mut.rate}{?}

\item{seq.length}{?}

\item{num.stgs}{?}

\item{selfing}{?}

\item{surv.matr}{?}

\item{repr.matr}{?}

\item{male.matr}{?}

\item{init.pop.sizes}{?}

\item{num.gen}{?}

\item{num.alleles}{?}

\item{allele.freqs}{?}
}
\description{
Initialize a landscape object
}

