% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.run.R
\name{rms.run}
\alias{rms.run}
\title{Run rmetasim}
\usage{
rms.run(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\value{
a modified \linkS4class{skeleSim.params} object with the results of
  a rmetasim run.
}
\description{
Run rmetasim
}

