% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.scenarioCheck.R
\name{rms.scenarioCheck}
\alias{rms.scenarioCheck}
\title{Check parameters for rmetasim}
\usage{
rms.scenarioCheck(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\description{
Check parameters for rmetasim
}
