% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.run.R
\name{rms.write}
\alias{rms.write}
\title{Write metasim file}
\usage{
rms.write(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\value{
Nothing
}
\description{
Write metasim landscape script to disk (one per scenario)
}
