% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSim.R
\name{runSim}
\alias{runSim}
\title{Run simulation}
\usage{
runSim(params, num.secs = NULL)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}

\item{num.secs}{number of seconds to run timing checks}
}
\description{
Run simulation
}
