\name{skeleSim-package}
\alias{skeleSim-package}
\alias{skeleSim}
\docType{package}
\title{
\packageTitle{skeleSim}
}
\description{
\packageDescription{skeleSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{skeleSim}
\packageIndices{skeleSim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{skeleSim}

Maintainer: \packageMaintainer{skeleSim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
