% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleSim.checks.R
\name{skeleSim.checks}
\alias{gen.scenario.check}
\alias{non.scenario.check}
\alias{overall.check}
\alias{skeleSim.checks}
\title{Check all simulation parameters}
\usage{
overall.check(params)

non.scenario.check(params)

gen.scenario.check(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\description{
Check all simulation parameters
}

