% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleSimGUI.R
\name{skeleSimGUI}
\alias{skeleSimGUI}
\title{GUI for skeleSim system}
\usage{
skeleSimGUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started.}
}
\value{
NULL

# markdown and shinyFiles import added for shiny app
}
\description{
This function starts the shiny simulation control panel
}

