% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funcs.R
\name{analysis_funcs}
\alias{analysis_funcs}
\title{Analysis functions}
\usage{
analysis_funcs(params)
}
\arguments{
\item{params}{Take results of a simulation from params@rep.sample and run Global, Locus, and Pairwise analyses}
}
\description{
Analyses for genind or multidna data
}

