% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.classes.R
\docType{class}
\name{fastsimcoal.params-class}
\alias{fastsimcoal.params}
\alias{fastsimcoal.params-class}
\title{fastsimcoal Parameters Class}
\description{
An S4 class storing parameters specific to fastsimcoal
}
\section{Slots}{

\describe{
\item{\code{fastsimcoal.exec}}{character string for the fastsimcoal command line
executable.}

\item{\code{pop.info}}{matrix of population sampling information created by the
\code{\link{fscPopInfo}} function.}

\item{\code{hist.ev}}{matrix of historical events created by the
\code{\link{fscHistEv}} function.}

\item{\code{locus.params}}{data.frame specifying loci to simulate created by the
\code{\link{fscLocusParams}} function.}
}}

