% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.run.R
\name{fsc.run}
\alias{fsc.run}
\title{Run fastsimcoal}
\usage{
fsc.run(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\value{
a modified \linkS4class{skeleSim.params} object with the results of
  a fastsimcoal run.
}
\description{
Run fastsimcoal
}
