% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.make.genind.R
\name{landscape.make.genind}
\alias{landscape.make.genind}
\title{Create genind object from landscape}
\usage{
landscape.make.genind(Rland, popnames = NULL)
}
\arguments{
\item{popnames}{character vector of population names.}

\item{l}{a \code{rmetasim} landscape object.}
}
\description{
Create genind object from landscape
}

