% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleSim.checks.R
\name{skeleSim.checks}
\alias{skeleSim.checks}
\alias{overall.check}
\alias{non.scenario.check}
\alias{gen.scenario.check}
\alias{analyses.check}
\title{Check all simulation parameters}
\usage{
overall.check(params)

non.scenario.check(params)

gen.scenario.check(params)

analyses.check(analyses.requested)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}

\item{analyses.requested}{A named logical vector with elements named Global, Locus, and Pairwise}
}
\description{
Check all simulation parameters
}
