% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleSim.internals.R
\name{skeleSim.internals}
\alias{currentLabel}
\alias{currentScenario}
\alias{skeleSim.internals}
\title{SkeleSim internal functions}
\usage{
currentScenario(params)

currentLabel(params)
}
\arguments{
\item{params}{a \linkS4class{skeleSim.params} object.}
}
\value{
\tabular{ll}{
    \code{currentScenario} \tab the parameters for the current scenario.\cr
    \code{currentLabel} \tab a character label representing current scenario
      and replicate.\cr
  }
}
\description{
SkeleSim internal functions
}
\keyword{internals}

